WIP - Extract actions from monolith file to individual action files; break down map action/reducers
