+++ app/src/actions.js
@@ -1,10 +1,3 @@
-// areas of interest
-export const SAVE_AREA_OF_INTEREST = 'SAVE_AREA_OF_INTEREST'; // Save the current working AoI as a "permanent" AoI
-export const UPDATE_WORKING_AREA_OF_INTEREST = 'UPDATE_WORKING_AREA_OF_INTEREST'; // Updates the current working AoI
-export const DELETE_AREA_OF_INTEREST = 'DELETE_AREA_OF_INTEREST';
-export const TOGGLE_AREA_OF_INTEREST_VISIBILITY = 'TOGGLE_AREA_OF_INTEREST_VISIBILITY';
-export const SET_RECENTLY_CREATED_AREA_OF_INTEREST = 'SET_RECENTLY_CREATED_AREA_OF_INTEREST';
-
 // map
 export const INIT_GOOGLE_MAPS = 'INIT_GOOGLE_MAPS';
 export const SET_LAYERS = 'SET_LAYERS';
@@ -13,7 +6,6 @@ export const SET_TILESET_URL = 'SET_TILESET_URL';
 export const SET_TILESET_ID = 'SET_TILESET_ID';
 export const SET_MAX_ZOOM = 'SET_MAX_ZOOM';
 export const SET_LAYER_OPACITY = 'SET_LAYER_OPACITY';
-export const SET_BASEMAP = 'SET_BASEMAP';
 export const SET_ZOOM = 'SET_ZOOM';
 export const SET_CENTER = 'SET_CENTER';
 export const SET_CENTER_TILE = 'SET_CENTER_TILE';
@@ -30,9 +22,6 @@ export const SET_SUBMENU = 'SET_SUBMENU';
 // modals
 export const SET_SUPPORT_MODAL_VISIBILITY = 'SET_SUPPORT_MODAL_VISIBILITY';
 export const SET_LAYER_MANAGEMENT_MODAL_VISIBILITY = 'SET_LAYER_MANAGEMENT_MODAL_VISIBILITY';
-export const SET_SEARCH_MODAL_VISIBILITY = 'SET_SEARCH_MODAL_VISIBILITY';
-export const SET_SEARCH_RESULTS_VISIBILITY = 'SET_SEARCH_RESULTS_VISIBILITY';
-export const SET_RECENT_VESSELS_VISIBILITY = 'SET_RECENT_VESSELS_VISIBILITY';
 export const SET_WELCOME_MODAL_VISIBILITY = 'SET_WELCOME_MODAL_VISIBILITY';
 export const SET_WELCOME_MODAL_CONTENT = 'SET_WELCOME_MODAL_CONTENT';
 export const SET_WELCOME_MODAL_CONTENT_ERROR = 'SET_WELCOME_MODAL_CONTENT_ERROR';
@@ -47,23 +36,6 @@ export const SET_LAYER_HUE = 'SET_LAYER_HUE';
 export const SET_WORKSPACE_LAYER_LABEL = 'SET_WORKSPACE_LAYER_LABEL';
 export const SHOW_CONFIRM_LAYER_REMOVAL_MESSAGE = 'SHOW_CONFIRM_LAYER_REMOVAL_MESSAGE';
 
-// layer library
-export const GET_LAYER_LIBRARY = 'GET_LAYER_LIBRARY';
-
-// Search
-export const SET_SEARCH_RESULTS = 'SET_SEARCH_RESULTS';
-export const SET_SEARCH_TERM = 'SET_SEARCH_TERM';
-export const SET_SEARCH_PAGE = 'SET_SEARCH_PAGE';
-export const SET_SEARCHING = 'SET_SEARCHING';
-
-// User actions
-export const SET_USER = 'SET_USER';
-export const SET_USER_PERMISSIONS = 'SET_USER_PERMISSIONS';
-export const SET_TOKEN = 'SET_TOKEN';
-export const LOGOUT = 'LOGOUT';
-export const TOKEN_SESSION = 'TOKEN_SESSION';
-export const SET_CURRENT_PATHNAME = 'SET_CURRENT_PATHNAME';
-
 // Filters actions
 export const SET_FLAG_FILTERS = 'SET_FLAG_FILTERS';
 export const SET_INNER_TIMELINE_DATES = 'SET_INNER_TIMELINE_DATES';
@@ -75,12 +47,7 @@ export const SET_OVERALL_TIMELINE_DATES = 'SET_OVERALL_TIMELINE_DATES';
 export const REWIND_TIMELINE = 'REWIND_TIMELINE';
 
 // Appearance actions
-export const SHARE_MODAL_OPEN = 'SHARE_MODAL_OPEN';
-export const SET_SHARE_MODAL_ERROR = 'SET_SHARE_MODAL_ERROR';
 export const SET_LAYER_INFO_MODAL = 'SET_LAYER_INFO_MODAL';
-
-// Contact actions
-export const FORM_RESPONSE = 'FORM_RESPONSE';
 
 // Vessel info actions
 export const ADD_VESSEL = 'ADD_VESSEL';
@@ -96,18 +63,6 @@ export const SET_PINNED_VESSEL_TITLE = 'SET_PINNED_VESSEL_TITLE';
 export const LOAD_PINNED_VESSEL = 'LOAD_PINNED_VESSEL';
 export const TOGGLE_PINNED_VESSEL_EDIT_MODE = 'TOGGLE_PINNED_VESSEL_EDIT_MODE';
 export const SET_PINNED_VESSEL_TRACK_VISIBILITY = 'SET_PINNED_VESSEL_TRACK_VISIBILITY';
-export const SET_RECENT_VESSEL_HISTORY = 'SET_RECENT_VESSEL_HISTORY';
-export const LOAD_RECENT_VESSEL_HISTORY = 'LOAD_RECENT_VESSEL_HISTORY';
-
-// report
-export const SHOW_POLYGON = 'SHOW_POLYGON';
-export const CLEAR_POLYGON = 'CLEAR_POLYGON';
-export const ADD_REPORT_POLYGON = 'ADD_REPORT_POLYGON';
-export const DELETE_REPORT_POLYGON = 'DELETE_REPORT_POLYGON';
-export const START_REPORT = 'START_REPORT';
-export const DISCARD_REPORT = 'DISCARD_REPORT';
-export const SET_REPORT_STATUS_SENT = 'SET_REPORT_STATUS_SENT';
-export const SET_REPORT_STATUS_ERROR = 'SET_REPORT_STATUS_ERROR';
 
 // heatmap
 export const INIT_HEATMAP_LAYERS = 'INIT_HEATMAP_LAYERS';
@@ -119,27 +74,5 @@ export const ADD_HEATMAP_LAYER = 'ADD_HEATMAP_LAYER';
 export const REMOVE_HEATMAP_LAYER = 'REMOVE_HEATMAP_LAYER';
 export const HIGHLIGHT_VESSELS = 'HIGHLIGHT_VESSELS';
 
-// upload
-export const CUSTOM_LAYER_UPLOAD_START = 'CUSTOM_LAYER_UPLOAD_START';
-export const CUSTOM_LAYER_UPLOAD_SUCCESS = 'CUSTOM_LAYER_UPLOAD_SUCCESS';
-export const CUSTOM_LAYER_UPLOAD_ERROR = 'CUSTOM_LAYER_UPLOAD_ERROR';
-
-
 // GA tracking, don't actually change the app state
-export const GA_SEARCH_RESULT_CLICKED = 'GA_SEARCH_RESULT_CLICKED';
-export const GA_VESSEL_POINT_CLICKED = 'GA_VESSEL_POINT_CLICKED';
-export const GA_MAP_POINT_CLICKED = 'GA_MAP_POINT_CLICKED';
-export const GA_PLAY_STATUS_TOGGLED = 'GA_PLAY_STATUS_TOGGLED';
-export const GA_OUTER_TIMELINE_DATES_UPDATED = 'GA_OUTER_TIMELINE_DATES_UPDATED';
-export const GA_EXTERNAL_LINK_CLICKED = 'GA_EXTERNAL_LINK_CLICKED';
-export const GA_DISCARD_REPORT = 'GA_DISCARD_REPORT';
-export const GA_MAP_CENTER_TILE = 'GA_MAP_CENTER_TILE';
-export const GA_INNER_TIMELINE_DATES_UPDATED = 'GA_INNER_TIMELINE_DATES_UPDATED';
-export const GA_INNER_TIMELINE_EXTENT_CHANGED = 'GA_INNER_TIMELINE_EXTENT_CHANGED';
-export const GA_SET_LAYER_OPACITY = 'GA_SET_LAYER_OPACITY';
 
-// Timebar
-export const SET_TIMEBAR_CHART_DATA = 'SET_TIMEBAR_CHART_DATA';
-
-// Literals
-export const LOAD_LITERALS = 'LOAD_LITERALS';
