+++ app/src/analytics/analyticsActions.js
@@ -1,18 +1,17 @@
 import debounce from 'lodash/debounce';
 import moment from 'moment';
 
-import {
-  GA_SEARCH_RESULT_CLICKED,
-  GA_VESSEL_POINT_CLICKED,
-  GA_MAP_POINT_CLICKED,
-  GA_EXTERNAL_LINK_CLICKED,
-  GA_DISCARD_REPORT,
-  GA_MAP_CENTER_TILE,
-  GA_OUTER_TIMELINE_DATES_UPDATED,
-  GA_INNER_TIMELINE_DATES_UPDATED,
-  GA_INNER_TIMELINE_EXTENT_CHANGED,
-  GA_SET_LAYER_OPACITY
-} from 'actions';
+export const GA_DISCARD_REPORT = 'GA_DISCARD_REPORT';
+export const GA_EXTERNAL_LINK_CLICKED = 'GA_EXTERNAL_LINK_CLICKED';
+export const GA_INNER_TIMELINE_DATES_UPDATED = 'GA_INNER_TIMELINE_DATES_UPDATED';
+export const GA_INNER_TIMELINE_EXTENT_CHANGED = 'GA_INNER_TIMELINE_EXTENT_CHANGED';
+export const GA_MAP_CENTER_TILE = 'GA_MAP_CENTER_TILE';
+export const GA_MAP_POINT_CLICKED = 'GA_MAP_POINT_CLICKED';
+export const GA_OUTER_TIMELINE_DATES_UPDATED = 'GA_OUTER_TIMELINE_DATES_UPDATED';
+export const GA_PLAY_STATUS_TOGGLED = 'GA_PLAY_STATUS_TOGGLED';
+export const GA_SEARCH_RESULT_CLICKED = 'GA_SEARCH_RESULT_CLICKED';
+export const GA_SET_LAYER_OPACITY = 'GA_SET_LAYER_OPACITY';
+export const GA_VESSEL_POINT_CLICKED = 'GA_VESSEL_POINT_CLICKED';
 
 /**
  * Only add here actions that are GA-exclusive.
