+++ app/src/analytics/analyticsMiddleware.js
@@ -1,27 +1,30 @@
 import {
   ADD_CUSTOM_LAYER,
-  GA_EXTERNAL_LINK_CLICKED,
-  GA_MAP_POINT_CLICKED,
-  GA_OUTER_TIMELINE_DATES_UPDATED,
-  GA_PLAY_STATUS_TOGGLED,
   ADD_REPORT_POLYGON,
   DELETE_REPORT_POLYGON,
-  GA_SEARCH_RESULT_CLICKED,
-  GA_VESSEL_POINT_CLICKED,
   SET_FLAG_FILTERS,
-  SET_SEARCH_TERM,
+  SET_REPORT_STATUS_SENT,
   SET_WORKSPACE_ID,
   SHOW_POLYGON,
   TOGGLE_LAYER_VISIBILITY,
   TOGGLE_LAYER_WORKSPACE_PRESENCE,
-  SET_REPORT_STATUS_SENT,
+  TOGGLE_VESSEL_PIN
+} from 'actions';
+import { SET_SEARCH_TERM } from 'search/searchActions';
+import {
   GA_DISCARD_REPORT,
-  GA_MAP_CENTER_TILE,
-  TOGGLE_VESSEL_PIN,
+  GA_EXTERNAL_LINK_CLICKED,
   GA_INNER_TIMELINE_DATES_UPDATED,
   GA_INNER_TIMELINE_EXTENT_CHANGED,
-  GA_SET_LAYER_OPACITY
-} from 'actions';
+  GA_MAP_CENTER_TILE,
+  GA_MAP_POINT_CLICKED,
+  GA_OUTER_TIMELINE_DATES_UPDATED,
+  GA_PLAY_STATUS_TOGGLED,
+  GA_SEARCH_RESULT_CLICKED,
+  GA_SET_LAYER_OPACITY,
+  GA_VESSEL_POINT_CLICKED
+} from 'analytics/analyticsActions';
+
 import isFunction from 'lodash/isFunction';
 import { FLAGS, SEARCH_QUERY_MINIMUM_LIMIT } from 'constants';
