+++ app/src/areasOfInterest/areasOfInterestActions.js
@@ -1,10 +1,8 @@
-import {
-  SAVE_AREA_OF_INTEREST,
-  UPDATE_WORKING_AREA_OF_INTEREST,
-  TOGGLE_AREA_OF_INTEREST_VISIBILITY,
-  SET_RECENTLY_CREATED_AREA_OF_INTEREST,
-  DELETE_AREA_OF_INTEREST
-} from 'actions';
+export const SAVE_AREA_OF_INTEREST = 'SAVE_AREA_OF_INTEREST'; // Save the current working AoI as a "permanent" AoI
+export const UPDATE_WORKING_AREA_OF_INTEREST = 'UPDATE_WORKING_AREA_OF_INTEREST'; // Updates the current working AoI
+export const DELETE_AREA_OF_INTEREST = 'DELETE_AREA_OF_INTEREST';
+export const TOGGLE_AREA_OF_INTEREST_VISIBILITY = 'TOGGLE_AREA_OF_INTEREST_VISIBILITY';
+export const SET_RECENTLY_CREATED_AREA_OF_INTEREST = 'SET_RECENTLY_CREATED_AREA_OF_INTEREST';
 
 export function saveAreaOfInterest() {
   return (dispatch, getState) => {
