+++ app/src/components/Map.jsx
@@ -11,14 +11,14 @@ import ControlPanel from 'containers/Map/ControlPanel';
 import Header from 'siteNav/containers/Header';
 import Timebar from 'timebar/containers/Timebar';
 import Modal from 'components/Shared/Modal';
-import Share from 'containers/Map/Share';
+import Share from 'share/containers/Share';
 import LayerInfo from 'layers/containers/LayerInfo';
 import ReportPanel from 'report/containers/ReportPanel';
 import MapLayers from 'containers/Layers/MapLayers';
 import LayerLibrary from 'layers/containers/LayerManagementModal';
 import SearchModal from 'search/containers/SearchModal';
 import SupportForm from 'siteNav/containers/SupportForm';
-import RecentVesselsModal from 'containers/Map/RecentVesselsModal';
+import RecentVesselsModal from 'recentVessels/containers/RecentVesselsModal';
 import WelcomeModal from 'containers/Map/WelcomeModal';
 import PromptLayerRemoval from 'containers/Map/PromptLayerRemoval';
 import NoLogin from 'containers/Map/NoLogin';
