+++ app/src/containers/Map/NoLogin.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import NoLogin from 'components/Map/NoLogin';
-import { login } from 'actions/user';
+import { login } from 'user/userActions';
 
 const mapStateToProps = () => ({});
