+++ app/src/search/searchReducer.js
@@ -1,6 +1,11 @@
 import {
-  SET_SEARCH_RESULTS, SET_SEARCH_TERM, SET_SEARCH_MODAL_VISIBILITY, SET_SEARCHING, SET_SEARCH_PAGE, SET_SEARCH_RESULTS_VISIBILITY
-} from 'actions';
+  SET_SEARCH_RESULTS,
+  SET_SEARCH_TERM,
+  SET_SEARCHING,
+  SET_SEARCH_PAGE,
+  SET_SEARCH_MODAL_VISIBILITY,
+  SET_SEARCH_RESULTS_VISIBILITY
+} from 'search/searchActions';
 
 const initialState = {
   entries: [],
