+++ app/src/share/shareActions.js
@@ -1,0 +1,30 @@
+export const SHARE_MODAL_OPEN = 'SHARE_MODAL_OPEN';
+export const SET_SHARE_MODAL_ERROR = 'SET_SHARE_MODAL_ERROR';
+
+/**
+ * Open or close the share modal
+ *
+ * @export openShareModal
+ * @param {boolean} open - true to open, false to close
+ * @returns {object}
+ */
+export function openShareModal(open) {
+  return {
+    type: SHARE_MODAL_OPEN,
+    payload: open
+  };
+}
+
+/**
+ * Set the error to display within the share modal
+ *
+ * @export setShareModalError
+ * @param {string} error - message to display
+ * @returns {object}
+ */
+export function setShareModalError(error) {
+  return {
+    type: SET_SHARE_MODAL_ERROR,
+    payload: error
+  };
+}
