+++ app/src/user/userReducer.js
@@ -1,4 +1,4 @@
-import { SET_USER_PERMISSIONS, SET_USER, SET_TOKEN, LOGOUT, SET_CURRENT_PATHNAME } from 'actions';
+import { SET_USER_PERMISSIONS, SET_USER, SET_TOKEN, LOGOUT } from 'user/userActions';
 
 const initialState = {
   userPermissions: null
@@ -14,8 +14,6 @@ export default function (state = initialState, action) {
       return Object.assign({}, state, { token: action.payload });
     case LOGOUT:
       return Object.assign({}, state, { token: null, loggedUser: null });
-    case SET_CURRENT_PATHNAME:
-      return Object.assign({}, state, { currentPathname: action.payload.pathname });
     default:
       return state;
   }
