+++ config/webpack.config.js
@@ -85,6 +85,9 @@ const webpackConfig = {
       siteNav: path.join(rootPath, 'app/src/siteNav'),
       timebar: path.join(rootPath, 'app/src/timebar'),
       analytics: path.join(rootPath, 'app/src/analytics'),
+      user: path.join(rootPath, 'app/src/user'),
+      recentVessels: path.join(rootPath, 'app/src/recentVessels'),
+      share: path.join(rootPath, 'app/src/share'),
       lib: path.join(rootPath, 'app/lib'),
       components: 'src/components',
       constants: 'src/constants.js',
